Moved DJSAutocomplete to its respective file and added the same default search query to mdn
